+++ flask-server/settings.py
@@ -1,0 +1,22 @@
+from flask import Flask
+from flask_migrate import Migrate
+from flask_login import LoginManager
+from flask_sqlalchemy import SQLAlchemy
+from database.credentials import USERNAME, PASSWORD, SECRET_KEY, DBNAME
+
+
+global app, db, migrate, login_manager
+app = Flask(__name__)
+
+# app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///test.db'
+app.config['SQLALCHEMY_DATABASE_URI'] = 'mysql+pymysql://' + \
+	USERNAME + ':' + PASSWORD + '@localhost/' + DBNAME
+app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
+app.config['SECRET_KEY'] = SECRET_KEY
+
+db = SQLAlchemy(app)
+migrate = Migrate(app, db)
+
+login_manager = LoginManager()
+login_manager.init_app(app)
+login_manager.login_view = 'login'
